requirements.txt
# RAG System Dependencies

# Core Framework
langchain>=0.3.0
langchain-google-genai>=2.0.0
langchain-groq>=0.3.0
langchain-community>=0.3.0

# Vector Store
chromadb>=0.5.0
langchain-chroma>=0.2.0

# Document Processing
pypdf>=4.0.0
unstructured>=0.15.0
beautifulsoup4>=4.12.0
requests>=2.31.0

# Embeddings (local alternative)
sentence-transformers>=3.0.0

# Environment Management
python-dotenv>=1.0.0

# Utilities
tqdm>=4.66.0
